// src/types/database.ts
export interface ProjectFile {
  path: string;
  content: string;
}